# docs/swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
        "/device_info/create": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Cria um novo dispositivo (base para a c\u00e2mera).",
                "operationId": "post_create_device",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    }
                ],
                "tags": [
                    "Manage Device Info"
                ]
            }
        },
        "/device_info/delete": {
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "delete_delete_device",
                "parameters": [
                    {
                        "description": "ID do dispositivo a ser deletado",
                        "name": "id",
                        "type": "string",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Manage Device Info"
                ]
            }
        },
...
        "/device_info/get_by_id": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Busca um dispositivo pelo ID.",
                "operationId": "get_get_device_by_id",
                "parameters": [
                    {
                        "description": "ID do dispositivo",
                        "name": "id",
                        "type": "string",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Manage Device Info"
                ]
            }
        },
        "/device_info/get_by_type": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Busca dispositivos por tipo.",
                "operationId": "get_get_device_by_type",
                "parameters": [
                    {
                        "description": "Tipo do dispositivo",
                        "name": "device_tipo",
                        "type": "string",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Manage Device Info"
                ]
            }
        },
        "/device_info/update": {
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Atualiza as informa\u00e7\u00f5es de um dispositivo.",
                "operationId": "put_update_device",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Device"
                        }
                    }
                ],
                "tags": [
                    "Manage Device Info"
                ]
            }
        },
...
    },
    "info": {
        "title": "API Video Streaming | Frame Processing | RTSP Management",
        "version": "1.0.0",
        "description": "API para gerenciamento de c\u00e2meras RTSP, Streaming de video e tratamento de FRAMES.",
        "contact": {
            "name": "Fabio Thomaz da Silva",
            "email": "fabio_thomaz@live.com",
            "url": null
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
...
    "definitions": {
...
        "Device": {
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "ID do registro"
                },
                "user_id": {
                    "type": "integer",
                    "description": "ID do usu\u00e1rio"
                },
                "device_id": {
                    "type": "integer",
                    "description": "ID do dispositivo (FK para device_info.id)"
                },
                "cam_name": {
                    "type": "string",
                    "description": "Nome da c\u00e2mera"
                },
                "grupo": {
                    "type": "string",
                    "description": "Grupo da c\u00e2mera (opcional)"
                },
                "full_cam_url_stream": {
                    "type": "string",
                    "description": "URL completa para acesso \u00e0 c\u00e2mera via stream"
                },
                "full_cam_url_rtsp": {
                    "type": "string",
                    "description": "URL completa para acesso \u00e0 c\u00e2mera via RTSP"
                },
                "cam_status": {
                    "type": "string",
                    "description": "Status da c\u00e2mera (online, offline, error)"
                },
                "device_config": {
                    "type": "object",
                    "description": "Configura\u00e7\u00f5es adicionais do dispositivo em JSON"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Data de cria\u00e7\u00e3o"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Data da \u00faltima atualiza\u00e7\u00e3o"
                }
            },
            "type": "object"
        },
...
    },
...
}